# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Sovereign Office of the Chief Justice & Trustee API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /trust/filings:
    get:
      operationId: listFilings
      tags: [filings]
      summary: List all trust filings
      responses:
        "200":
          description: List of filings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TrustFiling"
    post:
      operationId: createFiling
      tags: [filings]
      summary: Create a trust filing
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FilingInput"
      responses:
        "201":
          description: Filing created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrustFiling"
  /trust/filings/{id}:
    get:
      operationId: getFiling
      tags: [filings]
      summary: Get filing by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Filing
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrustFiling"
        "404":
          description: Not found
  /tribal/classify:
    post:
      operationId: classifyText
      tags: [classify]
      summary: Classify text through doctrine pipeline (creates NFR + task + calendar)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClassifyInput"
      responses:
        "201":
          description: Classification result with NFR, task, and calendar event
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClassifyResponse"
...
  /court/nfr/{id}:
    get:
      operationId: getNfr
      tags: [nfr]
      summary: Get NFR document by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: NFR document
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NfrDocument"
  /court/nfr/{id}/export-pdf:
    post:
      operationId: exportNfrPdf
      tags: [nfr]
      summary: Generate recorder-compliant PDF for an NFR document
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: PDF export result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PdfExportResult"
...
components:
  schemas:
...
    FilingInput:
      type: object
      properties:
        county: { type: string }
        state: { type: string }
        instrumentId: { type: integer }
        documentType: { type: string }
        notes: { type: string }
      required: [county, state]
...
    TrustFiling:
      type: object
      properties:
        id: { type: integer }
        instrumentId: { type: integer }
        county: { type: string }
        state: { type: string }
        filingStatus: { type: string }
        filingNumber: { type: ["string", "null"] }
        documentType: { type: ["string", "null"] }
        trustStatus: { type: ["string", "null"] }
        landClassification: { type: ["string", "null"] }
        notes: { type: ["string", "null"] }
        submittedAt: { type: ["string", "null"] }
        acceptedAt: { type: ["string", "null"] }
        rejectedAt: { type: ["string", "null"] }
        createdAt: { type: string }
        updatedAt: { type: string }
      required: [id, instrumentId, county, state, filingStatus, createdAt, updatedAt]
    ClassifyInput:
      type: object
      properties:
        text: { type: string }
      required: [text]
    ClassifyResponse:
      type: object
      properties:
        classification:
          type: object
          properties:
            id: { type: integer }
            actorType: { type: string }
            landStatus: { type: string }
            actionType: { type: string }
            rawText: { type: string }
        doctrine: { type: object }
        nfr: { type: object }
        task: { type: object }
        calendarEvent: { type: object }
    NfrDocument:
      type: object
      properties:
        id: { type: integer }
        classificationId: { type: integer }
        content: { type: string }
        pdfUrl: { type: ["string", "null"] }
        status: { type: string }
        createdAt: { type: string }
        updatedAt: { type: string }
      required: [id, classificationId, content, status, createdAt, updatedAt]
    PdfExportResult:
      type: object
      properties:
        success: { type: boolean }
        pdfUrl: { type: string }
        downloadUrl: { type: string }
        pages: { type: integer }
        checksum: { type: string }
        generatedAt: { type: string }
...