# lib/api-zod/src/generated/types/classifyInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Sovereign Office of the Chief Justice & Trustee API
 * OpenAPI spec version: 0.1.0
 */

export interface ClassifyInput {
  text: string;
}